refactor: setting e2e tests
